apps/katikaa-health-monitor/requirements.txt
# Core Application Dependencies
streamlit==1.28.1
pandas==2.1.1
plotly==5.17.0
numpy==1.24.3
requests==2.31.0

# Database Connectivity
mysql-connector-python==8.1.0
SQLAlchemy==2.0.21

# AWS Integration
boto3==1.28.17
botocore==1.31.17

# Environment Management
python-dotenv==1.0.0

# Data Processing
pytz==2023.3

# Visualization and Charts
matplotlib==3.7.2
seaborn==0.12.2

# HTTP and API
urllib3==2.0.4
certifi==2023.7.22

# Template Engine
jinja2==3.1.2

# Utilities
pydantic==2.4.1
typing-extensions==4.7.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Production Dependencies
gunicorn==21.2.0
uvicorn==0.23.2

# Monitoring
prometheus-client==0.17.1
